+++ src/moduls/daq/DiamondBoards/diamond.h
@@ -110,7 +110,7 @@ class TMdPrm : public TParamContr
 
 	//Attributes
 	TElem	pEl;		//Work atribute elements
-        int	&mTP,		//Board type
+        int64_t	&mTP,		//Board type
 		&mADDR,		//Board address
 		&mINT,		//Board interrupt
 		&mS_RATE,	//Sample rate
@@ -150,7 +150,7 @@ class TMdContr: public TController
 
         int64_t	period( )	{ return mPer; }
         string	cron( )		{ return mSched; }
-        int	prior( )	{ return mPrior; }
+        int64_t	prior( )	{ return mPrior; }
         int64_t lag( )		{ return mLag; }
 
 	AutoHD<TMdPrm> at( const string &nm )	{ return TController::at(nm); }
@@ -170,7 +170,7 @@ class TMdContr: public TController
 
 	//Attributes
 	Res	enRes;			//Resource for enable params
-	int	&mPrior;		//Process task priority
+	int64_t	&mPrior;		//Process task priority
 	TCfg	&mSched;		//Calc schedule
 	int64_t	mPer, mLag;
